+++ app/commands/view-feeds.js
@@ -1,4 +1,5 @@
 const Core = require("../../discord-bot-core");
+const Config = require("../config.json");
 
 module.exports = new Core.Command({
 	name: "view-feeds",
@@ -12,9 +13,8 @@ function invoke({ message, params, guildData, client }) {
 	if (!guildData)
 		return Promise.reject("Guild not setup");
 
-	const numToShow = 10;
-	const startIdx = params[0] ? (params[0] - 1) * numToShow : 0;
-	const endIdx = startIdx + numToShow + 1;
+	const startIdx = params[0] ? (params[0] - 1) * Config.viewFeedsPaginationLimit : 0;
+	const endIdx = startIdx + Config.viewFeedsPaginationLimit + 1;
 
 	let responseStr = guildData.feeds.map(f => f.toString()).slice(startIdx, endIdx).join("\n");
 	if (guildData.feeds.length > endIdx)